requirements.txt
# BioMCP API Server Requirements
# Core BioMCP dependencies
biomcp-python>=0.4.6

# API server dependencies
fastapi>=0.110.0
uvicorn>=0.34.2

# HTTP client for API examples
requests>=2.31.0

# LLM integration dependencies
httpx>=0.27.0
python-dotenv>=1.0.0

# Markdown to HTML conversion
markdown2>=2.4.0